fix: resolved data type inconsistencies in ensemble and flashcard CRUD
Addressed mismatched use of Long and Integer across controllers, repos, and models

+++ src/main/java/org/equipe4/flashpomobackend/repository/EnsembleRepository.java
@@ -14,6 +14,5 @@ import java.util.Optional;
  * @param <Integer> the type of the entity's primary key, in this case, Integer
  */
 public interface EnsembleRepository extends JpaRepository<Ensemble, Integer> {
-
-    Optional<Ensemble> findById(Long ensembleId);
+    Optional<Ensemble> findById(Integer ensembleId);
 }
